requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
disnake==2.5.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
mutagen==1.45.1
pycparser==2.21
pycryptodomex==3.14.1
PyNaCl==1.4.0
python-dotenv==0.20.0
pytube==12.1.0
requests==2.28.0
six==1.16.0
urllib3==1.26.9
websockets==10.3
yarl==1.7.2
yt-dlp==2022.5.18
#pytube>=12.1.0
git+https://github.com/Nobita-XD/p.git
